ci(SECREM-01): clear Rust 1.96 clippy lints (CI uses newer stable than local)
CI's `dtolnay/rust-toolchain@stable` resolves to 1.96.0, which enables lints
my local 1.94 didn't. Synced local to 1.96 and cleared the full set
(all mechanical, zero behavior change):

- 12× unnecessary_sort_by → sort_by_key (ascending) / sort_by_key + Reverse
  (descending) across sequencer, network/discovery, storage, marketplace.
- 2× unnecessary u32→u32 cast removed (q16 routing precompile).
- 1× useless .into_iter() in an IntoIterator arg (cli contract command).

`cargo clippy --workspace --all-targets -- -D warnings` → clean on 1.96.

+++ core/execution/src/precompiles/q16/routing.rs
@@ -870,12 +870,12 @@ mod tests {
         let input = make_canonical_input_uniform_input(Q16::ONE.0);
         let out = forward_decoded(&input);
         assert!(
-            (out.mentor_id as u32) < input.shape.output_dim,
+            out.mentor_id < input.shape.output_dim,
             "mentor_id {} must be < output_dim {}",
             out.mentor_id, input.shape.output_dim
         );
         assert!(
-            (out.adapter_id as u32) < input.shape.output_dim,
+            out.adapter_id < input.shape.output_dim,
             "adapter_id {} must be < output_dim {}",
             out.adapter_id, input.shape.output_dim
         );
